+++ src/onegov.core/onegov/core/orm/func.py
@@ -1,0 +1,7 @@
+from sqlalchemy.sql.functions import ReturnTypeFromArgs
+
+
+class unaccent(ReturnTypeFromArgs):
+    """ Produce a UNACCENT expression. """
+
+    pass